Separate CloudSync recovery protocol
Extract the persisted full-resync recovery state machine and barrier validation into a focused module while preserving the existing db-app API.
